Check in LICENSE and README
In preparation for publishing, I've added a README that briefly explains
the repo. License chosen is MIT -- explanation is in the README as well.

This is one of a few steps I'd like to perform before publishing on
Github for visibility. I also want to set up the scaffolding for cleanly
running Node and Rust programs off a single project root before then.