Solve Day 4: Camp Cleanup - Part 1
As always, algorithm is in the comments.

There was a very tricky gotcha that took me a few minutes to spot!
I had to explicitly cast bounds to numbers in parseRange, because I had
missed the fact that they come back strings initially.

This caused the comparison operators in doesContain to work in
lexicographic order instead of numeric order!

That one was a tough nut to crack for a few moments, heh.