Solve Day 7: No Space Left on Device - Parts 1 & 2
Okay, so this one, I actually solved the day it came out:

% ls -l day-07.js
-rw-r--r-- 1 ian ian 6579 Dec  7 16:28 day-07.js

But I did it under the influence of COVID (which also explains the
downtime for Advent of Code up until now!). If you'll notice, there are
some parts of the algorithm that I just *couldn't* for the life of me
write in an elegant way in the moment.

The parsing logic though, I'm impressed with myself, honestly. Even in
hindsight, the parse loop and the data structure feel quite elegant.

It's the traversal where it seems like my brain was coming up with
half-baked patterns and mixing them together. Like a printer that could
print out high quality images, but only for half the page before running
out of juice...

I was holding off all this time from moving forward with AoC partly
because I lost steam during the past few weeks where I've been ill, but
also partly because I was holding up this day's solution because I
wanted to refactor it.

Well, I *will* eventually refactor it, but when I was going through my
TODO and FIXME comments, I was amused! And I saw value in my ramblings
in the comments -- it shows how I verbalize the thought process behind
what I believe to be "elegance" in code.

So I'm checking all this in! I'll clean it up after I catch up on my
backlog.