Solve Day 8: Treetop Tree House - Part 1
As always, algorithm is in the comments.

Implementation side-winded through a lot of different iterations while I
was implementing, but I finally settled on generator functions for
getting matrix row and column iterables so I get the perfect combination
of readability and performance (lazy iteration, so we don't double
traverse, which we likely would have if we just naively built arrays for
rows and especially columns).